Refactoring payment instruments
